Added a new "allow_expand" setting
The new setting is a boolean that defaults to true.
It is only relevant when a large image is resized.
If true, the expand control is displayed just as before.
If false, the expand control is not displayed; the user
sees the scaled image but cannot zoom in to the original size.

I added this setting as an option to simplify the user interface.
Some users confuse the expand control with a close button, then
can't easily figure out how to un-expand the image and close.